⚡️::HealthKit 비동기처리
